created a __main__.py file for the project and moved the source to src folder
