+++ __main__.py
@@ -1,0 +1,7 @@
+import sys
+from src.final import MyWindow
+from PySide6.QtWidgets import QApplication
+
+my_app = QApplication([])
+my_win = MyWindow()
+sys.exit(my_app.exec())
